requirements.txt
discord.py>=2.3,<3
python-dotenv>=1.0,<2
Pillow>=10.0,<12